refactor: Restructure project directory layout
