+++ src/main/java/gift/domain/member/repository/MemberRepository.java
@@ -1,0 +1,13 @@
+package gift.domain.member.repository;
+
+import gift.domain.member.entity.Member;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByEmail(String email);
+
+}
