--- src/main/java/gift/member/repository/MemberRepository.java
@@ -1,13 +1,0 @@
-package gift.member.repository;
-
-import gift.member.entity.MemberEntity;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
-
-    Optional<MemberEntity> findByEmail(String email);
-
-}
